Create post-receive
post-receive builds a "dist" version of the application and runs it twice in two specific locations. Uses a 'panama canal' approach; one instance running on 9000 is shut down, then a new one on 9000 is started. Then the instance on 9001 is shut down, and a new one on 9001 is started. nginx is configured to load balance between 9000 and 9001 so there should be zero downtime. Note also that production.conf is not checked into source control; it contains sensitive information, a la http://www.playframework.com/documentation/latest/ApplicationSecret

note -e argument on first line; very important!

does simple quick check at the end to ensure correct number of JVMs are up (todo